исправления по 6 спринту  - автоматический тест проверки требует, что бы класс NODE был отдельным файлом, возвращаю.
